+++ nexus-ingestion-api/src/main/java/org/techbd/ingest/NexusIngestionApiApplication.java
@@ -1,4 +1,4 @@
-package org.techbd;
+package org.techbd.ingest;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
