+++ nexus-ingestion-api/src/main/java/org/techbd/ingest/commons/Constants.java
@@ -1,4 +1,4 @@
-package org.techbd.commons;
+package org.techbd.ingest.commons;
 
 public class Constants {
     public static final String SECRET_NAME = System.getenv("ORG_TECHBD_SERVICE_SECRET_NAME");
