+++ nexus-ingestion-api/src/main/java/org/techbd/ingest/processor/MessageProcessingStep.java
@@ -1,0 +1,8 @@
+package org.techbd.ingest.processor;
+
+// import org.techbd.ingest.model.RequestContext;
+// import org.springframework.web.multipart.MultipartFile;
+
+public interface MessageProcessingStep {
+    //void process(RequestContext context, MultipartFile file);
+}
